src/main/java/com/sectors/sectorsbackend/service/SectorService.java
package com.sectors.sectorsbackend.service;

import com.sectors.sectorsbackend.repository.SectorRepository;
import com.sectors.sectorsbackend.domain.Sector;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class SectorService {
    private final SectorRepository sectorRepository;

    @Autowired
    public SectorService(SectorRepository sectorRepository) {
        this.sectorRepository = sectorRepository;
    }

    public List<Sector> getAllSectors() {
        return sectorRepository.findAll();
    }

}
